% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best-efforts.R
\docType{data}
\name{best_efforts}
\alias{best_efforts}
\title{Strava activity best efforts}
\format{
Data frame with columns
\describe{
\item{id}{Activity ID}
\item{effort}{Effort description}
\item{start_index}{Effort start index in activity stream}
\item{end_index}{Effort end index in activity stream}
}
}
\source{
\href{https://developers.strava.com}{Strava API}
}
\usage{
data(best_efforts)
}
\description{
Data frame containing Strava activity best efforts.
}
\keyword{datasets}
