% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\docType{data}
\name{streams}
\alias{streams}
\title{Strava activity streams}
\format{
Data frame with columns
\describe{
\item{id}{Activity ID}
\item{distance}{Cumulative distance (metres)}
\item{time}{Cumulative time (seconds)}
\item{moving}{Whether moving}
\item{speed}{Smoothed speed (metres per second)}
\item{lat}{Latitude coordinate}
\item{lon}{Longitude coordinate}
\item{altitude}{Altitude (metres)}
\item{hr}{Heart rate (beats per minute)}
\item{cadence}{Cadence (rotations per minute)}
}
}
\source{
\href{https://developers.strava.com}{Strava API}
}
\usage{
data(streams)
}
\description{
Data frame containing Strava activity streams.
}
\keyword{datasets}
